% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clipboard}
\alias{clipboard}
\title{Copy an object in the clipboard}
\usage{
clipboard(x, sep = "\\t", row.names = FALSE, col.names = FALSE)
}
\arguments{
\item{x}{The object to copy.}

\item{sep}{A character to be used as separator for each column of the object}

\item{row.names}{Copy row names (default is FALSE)}

\item{col.names}{Copy column names (default is TRUE)}
}
\value{
copy the object as character in the clipboard
}
\description{
Copy an object in the clipboard (needs xclip to be installed)
}
\author{
freecube source:http://stackoverflow.com/questions/10959521/how-to-write-to-clipboard-on-ubuntu-linux-in-r
}
