% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_operations.R
\name{expandRows}
\alias{expandRows}
\title{expands rows - i dont remember what it was for.}
\usage{
expandRows(dataset, count, count_is_col = TRUE)
}
\arguments{
\item{dataset}{a data frame.}

\item{count}{numeric. Expand rows by how many.}

\item{count_is_col}{boolean.}
}
\value{
A data.frame with extended rows from \code{x}.
}
\description{
expands rows - i dont remember what it was for.
}
\examples{
#ExpandRows(???)
#??
}
