% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pallette.R
\name{get_palette_from_theme}
\alias{get_palette_from_theme}
\title{Get palette from current theme}
\usage{
get_palette_from_theme()
}
\value{
A character string naming the palette.
}
\description{
Attempts to retrieve the .palette element from the current ggplot2 theme.
Defaults to "adj_colorblind" if not found.
}
