% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxtras.R
\name{ggplot2_settings_on}
\alias{ggplot2_settings_on}
\title{Set options for ggplot2}
\usage{
ggplot2_settings_on()
}
\description{
Set default ggplot2 theme to theme_science and default discrete scales to science palettes.
}
