% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_loaders.R
\name{loadWigs}
\alias{loadWigs}
\title{Loads wig files.}
\usage{
loadWigs(filename, filepath)
}
\arguments{
\item{filename}{A name of your file.}

\item{filepath}{Path to file.}
}
\value{
A loaded wig file.
}
\description{
Loads wig files.
}
\examples{
# load just one file
#library(rtracklayer)
#loadWigs("tobeloaded.wig", "path/to/file")

# load all files from location
#fpath <- "/export/kjempetujafs/valen-group/projects/micropeptides/data/Ribo-Seq/Dome/fwd/fixed"
#forward <- list.files(fpath)
#FLOSSlengths <- sapply(forward, function(x) as.numeric(gsub("[^0-9]", "", x)), USE.NAMES = F) #lengths of footprints from names of files
#forwardFP <- lapply(forward, loadWigs, filepath = fpath)
}
