% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pallette.R
\docType{data}
\name{pastel}
\alias{pastel}
\title{Pastel color palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
pastel
}
\description{
Pastel color palette
}
\keyword{datasets}
