% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{plotRanges}
\alias{plotRanges}
\title{Plot genomic ranges from the IRanges object or GRanges object.}
\usage{
plotRanges(
  x,
  xlim = x,
  main = deparse(substitute(x)),
  col = "black",
  sep = 0.5,
  ...
)
}
\arguments{
\item{x}{a IRanges object.}

\item{xlim}{like in plot.}

\item{main}{A title of the plot.}

\item{col}{color. Default "black". e.g. "blue".}

\item{sep}{Default 0.5.}

\item{...}{Other params you can pass to plot function.}
}
\value{
A plot.
}
\description{
Plot genomic ranges from the IRanges object or GRanges object.
}
\examples{
library(IRanges)
plotRanges(IRanges(1:10, width=10:1, names=letters[1:10]))
}
