% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pallette.R
\docType{data}
\name{resurrect}
\alias{resurrect}
\title{Resurrect color palette}
\format{
An object of class \code{character} of length 64.
}
\usage{
resurrect
}
\description{
Resurrect color palette
}
\keyword{datasets}
