% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxtras.R
\docType{package}
\name{rxtras}
\alias{rxtras-package}
\alias{rxtras}
\title{rxtras package}
\description{
Extra functions and addins that you might find useful.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JokingHero/rxtras}
  \item Report bugs at \url{https://github.com/JokingHero/rxtras/issues}
}

}
