% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_scale.R
\name{scale_gradient}
\alias{scale_gradient}
\title{Scale gradient for ggplot2}
\usage{
scale_gradient(
  min,
  max,
  aesthetics = "fill",
  gradient = c("#D62929", "#ffffff", "#2982D6"),
  na.value = "#dddddd"
)
}
\arguments{
\item{min}{Minimal value in the data}

\item{max}{Maximal value in the data}

\item{aesthetics}{"colour" or "fill" or both}

\item{gradient}{A vector of 3 colours: low, middle and high}

\item{na.value}{What colour to put on the na.value}
}
\value{
a scale_fill_gradient
}
\description{
Depending on min and max will apply gradient to the plot
if min < 0, use 3 colours and set limits to -max_val, +max_val
else use 2 colours and set limits to 0, max

Alternative dark red-blue: c("#750e13", "#ffffff", "#003a6d")
Alternative violet-green: c("#491d8b", "#ffffff", "#004144")
}
