% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_science}
\alias{scale_colour_science}
\alias{scale_color_science}
\alias{scale_fill_science}
\title{Science Discrete Color Scales}
\usage{
scale_colour_science(palette = NULL, ...)

scale_color_science(palette = NULL, ...)

scale_fill_science(palette = NULL, ...)
}
\arguments{
\item{palette}{Name of the palette. If NULL, uses the palette from the current theme.}

\item{...}{other arguments passed to discrete_scale}
}
\description{
Science Discrete Color Scales
}
