% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_science_div}
\alias{scale_colour_science_div}
\alias{scale_color_science_div}
\alias{scale_fill_science_div}
\title{Science Diverging Color Scale}
\usage{
scale_colour_science_div(palette = NULL, ...)

scale_color_science_div(palette = NULL, ...)

scale_fill_science_div(palette = NULL, ...)
}
\arguments{
\item{palette}{Name of the palette.}

\item{...}{other arguments passed to scale_colour_gradient2}
}
\description{
Science Diverging Color Scale
}
