% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_science_seq}
\alias{scale_colour_science_seq}
\alias{scale_color_science_seq}
\alias{scale_fill_science_seq}
\title{Science Sequential Color Scale}
\usage{
scale_colour_science_seq(palette = "adj_colorblind", ...)

scale_color_science_seq(palette = "adj_colorblind", ...)

scale_fill_science_seq(palette = "adj_colorblind", ...)
}
\arguments{
\item{palette}{Name of the palette.}

\item{...}{other arguments passed to scale_colour_gradient}
}
\description{
Science Sequential Color Scale
}
