% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pallette.R
\docType{data}
\name{science_colors}
\alias{science_colors}
\title{Science Palettes Colors}
\format{
An object of class \code{list} of length 3.
}
\usage{
science_colors
}
\description{
Define the color vectors for the science theme.
}
\keyword{datasets}
