% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pallette.R
\docType{data}
\name{science_gradient_colors}
\alias{science_gradient_colors}
\title{Science Gradient Colors}
\format{
An object of class \code{list} of length 3.
}
\usage{
science_gradient_colors
}
\description{
Define the low, mid, and high colors for science gradients.
}
\keyword{datasets}
