% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pallette.R
\docType{data}
\name{science_gradients}
\alias{science_gradients}
\title{Alias for science_gradient_colors}
\format{
An object of class \code{list} of length 3.
}
\usage{
science_gradients
}
\description{
Alias for science_gradient_colors for compatibility.
}
\keyword{datasets}
