% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pallette.R
\name{science_pal}
\alias{science_pal}
\title{Science Palette function for discrete_scale}
\usage{
science_pal(palette = "adj_colorblind")
}
\arguments{
\item{palette}{Name of the palette.}
}
\value{
A function that takes an integer n and returns n colors.
}
\description{
Returns a function that generates a vector of colors.
}
