% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pallette.R
\name{science_palette}
\alias{science_palette}
\title{Science Palette Generator}
\usage{
science_palette(n, palette = NULL)
}
\arguments{
\item{n}{Number of colors to return.}

\item{palette}{Name of the palette ("adj_colorblind", "pastel", "resurrect").}
}
\value{
A character vector of hex colors.
}
\description{
Returns a vector of colors for a given palette name and number of colors.
}
