% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{substrRight}
\alias{substrRight}
\title{Extract substrings in a character vector counting from right.}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{a character vector.}

\item{n}{integer. The first element to be replaced.}
}
\value{
A substracted character from \code{x}.
}
\description{
Extract substrings in a character vector counting from right.
}
\examples{
substrRight("ABCD", 2)
#CD

}
