% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_science}
\alias{theme_science}
\title{Set theme for ggplot2}
\usage{
theme_science(base_size = 16)
}
\arguments{
\item{base_size}{Font size}
}
\value{
a ggplot2 theme
}
\description{
The theme set supports 10 colors in the colorblind friendly, yet
beautiful pastille colors, and afterwards it backs out to 64 colors from
resurrect palette. Theme is based on `ggthemes::theme_pander()`.
}
\author{
See ggtheme package
}
