% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_science_palette}
\alias{theme_science_palette}
\title{Set theme for ggplot2 (Science Style)}
\usage{
theme_science_palette(palette = "adj_colorblind", ...)
}
\arguments{
\item{palette}{Science palette to use ("adj_colorblind", "pastel", "resurrect")}

\item{...}{Additional arguments passed to theme_science()}
}
\description{
Set theme for ggplot2 (Science Style)
}
