% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO_Addin.R
\name{TODO}
\alias{TODO}
\title{Find all TODO comments in a current Project}
\usage{
TODO()
}
\description{
Call this as an addin to find all comments with TODO, make list of them,
click on them ponts to clicked TODO location.
}
