% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pallette.R
\name{validate_palette}
\alias{validate_palette}
\title{Validate palette name}
\usage{
validate_palette(palette)
}
\arguments{
\item{palette}{Palette name to validate.}
}
\value{
The palette name if valid, throws error otherwise.
}
\description{
Validate palette name
}
