% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{as.pngpdf}
\alias{as.pngpdf}
\title{Wrapper around gglplot2::ggsave with presets.}
\usage{
as.pngpdf(
  filename,
  plot = ggplot2::last_plot(),
  scale = 1,
  width = 12,
  height = 8,
  units = "in",
  dpi = 400
)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{scale}{Multiplicative scaling factor.}

\item{width, height, units}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}
}
\value{
A plot saved as .png and .pdf, returns nothing.
}
\description{
Wrapper around gglplot2::ggsave with presets.
}
