% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{cpaste}
\alias{cpaste}
\title{Paste an object from the clipboard}
\usage{
cpaste(header = FALSE, ...)
}
\arguments{
\item{header}{Is there a header?}

\item{...}{What arguments to pass to the read.delim function?}
}
\value{
copy the object as character in the clipboard
}
\description{
Paste an object from the clipboard (needs xsel to be installed)
}
\author{
isomorphismes source:http://stackoverflow.com/questions/13438556/how-do-i-copy-and-paste-data-into-r-from-the-clipboard
}
