% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_science}
\alias{theme_science}
\title{Set theme for ggplot2 (Science Style)}
\usage{
theme_science(
  base_size = 12,
  base_family = "sans",
  nomargin = FALSE,
  fc = "black",
  gM = TRUE,
  gm = TRUE,
  gc = "grey",
  gl = "dashed",
  boxes = TRUE,
  bc = "white",
  pc = "transparent",
  lp = "top",
  axis = 1
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{nomargin}{If true will remove all margins.}

\item{fc}{text colour, default is black}

\item{gM}{grid Major}

\item{gm}{grid minor}

\item{gc}{grid color}

\item{gl}{grid linetype}

\item{boxes}{whether to keep boxes around the plot}

\item{bc}{background color, default white}

\item{pc}{panel colour background, default transparent}

\item{lp}{legend position, default is top}

\item{axis}{0, 1, 2, 3 - default is 1}
}
\description{
Set theme for ggplot2 (Science Style)
}
